move processedAnnotationData to Base
