+++ deet/data_models/eppi.py
@@ -1,20 +1,17 @@
 """EPPI-specific data models extending the core models."""
 
-import csv
 from collections.abc import Callable
-from pathlib import Path
-from typing import Any, Literal
+from typing import Any
 
-from loguru import logger
 from pydantic import BaseModel, ConfigDict, Field, model_validator
 from pydantic.alias_generators import to_camel
 
 from deet.data_models.base import (
-    AnnotationType,
     Attribute,
     AttributeType,
     Document,
     GoldStandardAnnotation,
+    ProcessedAnnotationData,
 )
 
 
@@ -200,185 +197,15 @@ class EppiRawData(BaseModel):
         return all_attributes
 
 
-class ProcessedAnnotationData(BaseModel):
+class ProcessedEppiAnnotationData(ProcessedAnnotationData):
     """
-    Structured result from annotation processing.
+    Structured result from EPPI annotation processing.
 
-    This model provides a clean, validated structure for all processed
-    annotation data with useful properties and methods.
+    This differs from Base ProcessedAnnotationData by specifying raw_data as an
+    EppiRawData object
     """
 
-    attributes: list[EppiAttribute]
-    documents: list[EppiDocument]
-    annotations: list[EppiGoldStandardAnnotation]
-    annotated_documents: list[EppiGoldStandardAnnotatedDocument]
-    attribute_id_to_label: dict[int, str]
     raw_data: EppiRawData
-
-    def _custom_prompts_cli(self) -> None:
-        """
-        Use an interactive CLI to have the user enter custom prompts.
-
-        Args:
-            attribute (Attribute): a single (Eppi)Attribute
-
-        """
-        for attribute in self.attributes:
-            attribute.enter_custom_prompt()
-
-    def export_attributes_csv_file(self, filepath: Path) -> None:
-        """
-        Write a csv file containing all attributes for prompt population.
-
-        Args:
-            filepath (Path): outfile path.
-
-
-        """
-        if filepath.suffix != ".csv":
-            bad_filetype = "file ending must be .csv"
-            raise ValueError(bad_filetype)
-        for attribute in self.attributes:
-            attribute.write_to_csv(filepath=filepath)
-
-        logger.info(f"wrote attributes to file {filepath}.")
-
-    def _import_prompts_csv_file(
-        self, filepath: Path, *, overwrite: bool = True
-    ) -> None:
-        """
-        Import prompts from a csv file.
-
-        Args:
-            filepath (Path): attribute/prompt input file.
-            overwrite (bool, optional): Overwrite existing prompts. Defaults to True.
-
-        """
-        if not filepath.exists():
-            no_file = f"CSV file not found: {filepath}"
-            raise FileNotFoundError(no_file)
-
-        if filepath.suffix != ".csv":
-            bad_suffix = "File must have .csv extension"
-            raise ValueError(bad_suffix)
-
-        with filepath.open(mode="r", newline="", encoding="utf-8") as f:
-            reader = csv.DictReader(f)
-
-            if reader.fieldnames is None:
-                empty_csv = "CSV file is empty or has no headers"
-                raise ValueError(empty_csv)
-
-            required_fields = ["attribute_id", "prompt"]
-            for field in required_fields:
-                if field not in reader.fieldnames:
-                    csv_missing_fields = (
-                        f"CSV must contain '{field}' column. "
-                        f"Found columns: {', '.join(reader.fieldnames)}"
-                    )
-
-                    raise ValueError(csv_missing_fields)
-
-            rows_processed = 0
-            for row in reader:
-                # find attribute_id match
-                attribute_id = int(row["attribute_id"])
-                matching_attribute = None
-
-                for attribute in self.attributes:
-                    if attribute.attribute_id == attribute_id:
-                        matching_attribute = attribute
-                        break
-
-                if matching_attribute is None:
-                    logger.warning(
-                        f"No attribute found with ID {attribute_id}, skipping row"
-                    )
-                    continue
-
-                # populate prompt using the Attribute method
-                try:
-                    matching_attribute.populate_prompt_from_dict(
-                        row, overwrite=overwrite
-                    )
-                    rows_processed += 1
-                except ValueError as e:
-                    logger.error(
-                        f"Error processing row for attribute {attribute_id}: {e}"
-                    )
-
-            logger.info(f"Processed {rows_processed} prompts from {filepath}")
-
-    def populate_custom_prompts(
-        self, method: Literal["cli", "file"], filepath: Path | None = None, **kwargs
-    ) -> None:
-        """
-        Populate custom prompts.
-
-        Args:
-            method (Literal["cli", "file"])
-            filepath (Path | None): infile path.
-
-        Raises:
-            FileNotFoundError: if method is file and there's no filepath.
-
-        """
-        if method == "cli":
-            self._custom_prompts_cli()
-        elif method == "file":
-            if filepath is None:
-                missing_filepath = "please specify a filepath!"
-                raise FileNotFoundError(missing_filepath)
-            self._import_prompts_csv_file(filepath=filepath, **kwargs)
-        else:
-            not_impl = f"method {method} is not implemented. use cli or file."
-            raise NotImplementedError(not_impl)
-
-    @property
-    def total_attributes(self) -> int:
-        """Total number of attributes processed."""
-        return len(self.attributes)
-
-    @property
-    def total_documents(self) -> int:
-        """Total number of documents processed."""
-        return len(self.documents)
-
-    @property
-    def total_annotations(self) -> int:
-        """Total number of annotations processed."""
-        return len(self.annotations)
-
-    @property
-    def total_annotated_documents(self) -> int:
-        """Total number of documents with annotations."""
-        return len(self.annotated_documents)
-
-    def get_attributes_by_type(self, attribute_type: str) -> list[EppiAttribute]:
-        """Get all attributes of a specific type."""
-        return [
-            attr for attr in self.attributes if attr.attribute_type == attribute_type
-        ]
-
-    def get_documents_with_annotations(self) -> list[EppiDocument]:
-        """Get only documents that have annotations."""
-        annotated_doc_ids = {doc.document_id for doc in self.annotated_documents}
-        return [doc for doc in self.documents if doc.document_id in annotated_doc_ids]
-
-    def get_annotations_by_type(
-        self, annotation_type: AnnotationType
-    ) -> list[EppiGoldStandardAnnotation]:
-        """Get all annotations of a specific type (human/llm)."""
-        return [
-            ann for ann in self.annotations if ann.annotation_type == annotation_type
-        ]
-
-    def get_attribute_by_id(self, attribute_id: int) -> EppiAttribute | None:
-        """Get an attribute by its ID."""
-        for attr in self.attributes:
-            if attr.attribute_id == attribute_id:
-                return attr
-        return None
 
 
 class AttributeAnswerCoT(BaseModel):
